+++ src/test/kotlin/domain/lotto/LottoNumberTest.kt
@@ -1,6 +1,5 @@
-package domain
+package domain.lotto
 
-import domain.lotto.LottoNumber
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.params.ParameterizedTest
