+++ src/test/kotlin/domain/lotto/LottoTest.kt
@@ -1,7 +1,5 @@
-package domain
+package domain.lotto
 
-import domain.lotto.Lotto
-import domain.lotto.LottoNumber
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
